requirements.txt
streamlit
pandas
plotly
snowflake-connector-python
python-dotenv
scikit-learn
